[ReLayout] Small optimization in looping through children.
When looping through children, we always set the absolute child's
nextChild property, but that always gets overriden when observing the
next absolute child in the list. The only reason we set it was for the
last absolute child, so we should just check the *last* absolute child
after we're done looping.