[ReLayout] Fix reference equality issue with bytecode mode.
Summary:Reference equality is tricky. What we had worked well for
ocamlopt, but that's because an optimization pass had prevented a
reallocation of the dummyMeasure function. In ocamlc, we did not enjoy
the same optimization so the reference equality was failing. Made it
explicit now with an option. This is fine because most of the time,
the reference is `None`, and seldom do we add any measure functions.

Test Plan:

Reviewers:

CC: