[ReLayout] Make style mutable.
Summary:
This significantly improves performance for external bindings which do
imperative setHeight setWidth because it prevents us from having to copy
the entire style on each mutation.

It shouldn't hurt the performance of declarative use cases who don't
mutate the style. The only possible way it could hurt that use case is
if we were relying on the immutability to be able to allocate a layout
node without having to do {...defaultLayout, blah: 0}. When no style is
ever set, the immutable case is far better because it requires zero
allocations to use the default style. However, almost every layout node
*should* set some style, so it's worth it. If that ends up not being
true, we can still use a `dummyStyle` marker for "not mutated style" and
then be very careul in all the bindings to check for `=== dummyStyle`
before setting.

Test Plan:

Reviewers:

CC: